Refactor the webhook checking for fluid pod, To #43000843
