+++ pkg/utils/fluid_test.go
@@ -1,0 +1,84 @@
+/*
+Copyright 2023 The Fluid Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package utils
+
+import (
+	"testing"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func TestIsFluidPod(t *testing.T) {
+	type args struct {
+		pod *corev1.Pod
+	}
+	tests := []struct {
+		name string
+		args args
+		want bool
+	}{
+		{
+			name: "Data load Pod with correct role label is recognized as Fluid Pod",
+			args: args{
+				pod: &corev1.Pod{
+					ObjectMeta: metav1.ObjectMeta{
+						Labels: map[string]string{
+							"role": "dataload-pod",
+							"app":  "alluxio",
+						},
+					},
+				},
+			},
+			want: true,
+		},
+		{
+			name: "Pod with Fluid Pod label but incorrect role label is recognized as Fluid Pod",
+			args: args{
+				pod: &corev1.Pod{
+					ObjectMeta: metav1.ObjectMeta{
+						Labels: map[string]string{
+							"role": "web-pod",
+							"app":  "jindofs",
+						},
+					},
+				},
+			},
+			want: true,
+		},
+		{
+			name: "Pod with no Fluid Pod labels is not recognized as Fluid Pod",
+			args: args{
+				pod: &corev1.Pod{
+					ObjectMeta: metav1.ObjectMeta{
+						Labels: map[string]string{
+							"role": "web-pod",
+						},
+					},
+				},
+			},
+			want: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := IsPodManagedByFluid(tt.args.pod); got != tt.want {
+				t.Errorf("IsPodManagedByFluid() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
