+++ pkg/webhook/scheduler/mutating/schedule_pod_handler_test.go
@@ -1169,63 +1169,3 @@ func TestAddScheduleInfoToPodWithReferencedDataset(t *testing.T) {
 		}
 	}
 }
-
-func TestIsFluidPod(t *testing.T) {
-	type args struct {
-		pod *corev1.Pod
-	}
-	tests := []struct {
-		name string
-		args args
-		want bool
-	}{
-		{
-			name: "Data load Pod with correct role label is recognized as Fluid Pod",
-			args: args{
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"role": "dataload-pod",
-							"app":  "alluxio",
-						},
-					},
-				},
-			},
-			want: true,
-		},
-		{
-			name: "Pod with Fluid Pod label but incorrect role label is recognized as Fluid Pod",
-			args: args{
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"role": "web-pod",
-							"app":  "jindofs",
-						},
-					},
-				},
-			},
-			want: true,
-		},
-		{
-			name: "Pod with no Fluid Pod labels is not recognized as Fluid Pod",
-			args: args{
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"role": "web-pod",
-						},
-					},
-				},
-			},
-			want: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := isFluidPod(tt.args.pod); got != tt.want {
-				t.Errorf("IsFluidPod() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
